requirements.txt
loguru>=0.7.0
jsonpath>=0.82.2
requests>=2.31.0
pywin32>=306
PyYAML>=6.0.1
keyboard>=0.13.5
aiohttp>=3.9.3